move util to base64.py
